refactor(menu): rename scss files to use .module.scss
