+++ draft-packages/menu/KaizenDraft/Menu/MenuDropdown.module.scss
@@ -1,0 +1,27 @@
+@import "~@kaizen/component-library/styles/layers";
+@import "~@kaizen/design-tokens/sass/border";
+@import "~@kaizen/design-tokens/sass/color";
+@import "~@kaizen/design-tokens/sass/shadow";
+@import "~@kaizen/design-tokens/sass/spacing";
+@import "./variables";
+
+$menu-container-width: 248px;
+
+.menuContainer {
+  z-index: $ca-z-index-dropdown;
+  box-sizing: border-box;
+  background: $color-white;
+  color: $color-purple-800;
+  border-radius: $border-solid-border-radius;
+  box-shadow: $shadow-large-box-shadow;
+  padding: $spacing-xs 0;
+  text-align: left;
+
+  [dir="rtl"] & {
+    text-align: right;
+  }
+}
+
+.defaultWidth {
+  width: $menu-container-width;
+}
