+++ draft-packages/menu/KaizenDraft/Menu/StatelessMenu.module.scss
@@ -1,0 +1,3 @@
+.buttonWrapper {
+  display: inline-block;
+}
