+++ draft-packages/menu/KaizenDraft/Menu/_variables.scss
@@ -1,0 +1,3 @@
+@import "~@kaizen/design-tokens/sass/spacing";
+
+$side-padding: $spacing-sm;
