+++ draft-packages/menu/KaizenDraft/Menu/components/MenuItem.module.scss
@@ -1,67 +1,8 @@
-@import "~@kaizen/design-tokens/sass/spacing";
-@import "~@kaizen/design-tokens/sass/shadow";
-@import "~@kaizen/design-tokens/sass/border";
-@import "~@kaizen/design-tokens/sass/color";
-@import "~@kaizen/design-tokens/sass/typography";
-@import "~@kaizen/component-library/styles/layers";
 @import "~@kaizen/deprecated-component-library-helpers/styles/layout";
-
-$menu-container-width: 248px;
-$side-padding: $spacing-sm;
-$gutter: calc(#{$spacing-md} * 1 / 6);
-$padding: $spacing-xs;
-
-.buttonWrapper {
-  display: inline-block;
-}
-
-.menuContainer {
-  z-index: $ca-z-index-dropdown;
-  box-sizing: border-box;
-  background: $color-white;
-  color: $color-purple-800;
-  border-radius: $border-solid-border-radius;
-  box-shadow: $shadow-large-box-shadow;
-  padding: $spacing-xs 0;
-  text-align: left;
-
-  [dir="rtl"] & {
-    text-align: right;
-  }
-}
-
-.defaultWidth {
-  width: $menu-container-width;
-}
-
-.menuSection {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-
-  &:not(:last-of-type) {
-    &::after {
-      width: 100%;
-      height: 1px;
-      background-color: $border-solid-border-color;
-      content: "";
-      display: block;
-      margin: 5px 0;
-    }
-  }
-}
-
-.header {
-  padding: 10px $side-padding 5px;
-  color: $color-purple-800;
-  margin: 0 $gutter;
-  font-family: $typography-heading-6-font-family;
-  font-weight: $typography-heading-6-font-weight;
-  font-size: $typography-heading-6-font-size;
-  line-height: $typography-heading-6-line-height;
-  letter-spacing: $typography-heading-6-letter-spacing;
-  display: block;
-}
+@import "~@kaizen/design-tokens/sass/color";
+@import "~@kaizen/design-tokens/sass/spacing";
+@import "~@kaizen/design-tokens/sass/typography";
+@import "../variables";
 
 .menuListItem {
   display: flex;
@@ -76,7 +17,7 @@ $padding: $spacing-xs;
   border: 2px transparent solid;
   text-align: left;
   padding: 6px calc(#{$side-padding} - 2px);
-  margin: 0 $padding;
+  margin: 0 $spacing-xs;
   min-height: calc(1.75 * #{$spacing-md});
   border-radius: 4px;
   font-family: $typography-paragraph-body-font-family;
