+++ draft-packages/menu/KaizenDraft/Menu/components/MenuList.module.scss
@@ -1,0 +1,36 @@
+@import "~@kaizen/design-tokens/sass/spacing";
+@import "~@kaizen/design-tokens/sass/border";
+@import "~@kaizen/design-tokens/sass/color";
+@import "~@kaizen/design-tokens/sass/typography";
+@import "../variables";
+
+$gutter: calc(#{$spacing-md} * 1 / 6);
+
+.header {
+  padding: 10px $side-padding 5px;
+  color: $color-purple-800;
+  margin: 0 $gutter;
+  font-family: $typography-heading-6-font-family;
+  font-weight: $typography-heading-6-font-weight;
+  font-size: $typography-heading-6-font-size;
+  line-height: $typography-heading-6-line-height;
+  letter-spacing: $typography-heading-6-letter-spacing;
+  display: block;
+}
+
+.menuSection {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+
+  &:not(:last-of-type) {
+    &::after {
+      width: 100%;
+      height: 1px;
+      background-color: $border-solid-border-color;
+      content: "";
+      display: block;
+      margin: 5px 0;
+    }
+  }
+}
